Obj collisions
Also moved obj movement code to its own file for neatness
